Add test api-promise/var-declarator-promise
