+++ src/transforms/v2-to-v3/__fixtures__/api-promise/global-import.output.js
@@ -4,14 +4,3 @@ const client = new DynamoDB();
 
 // ExpressionStatement
 client.listTables();
-
-// Promise call on request in variable declarator
-const listTablesRequest = client.listTables();
-listTablesRequest
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
-
-const listTagsOfResourceRequest = client.listTagsOfResource({ ResourceArn: "STRING_VALUE" });
-listTagsOfResourceRequest
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
