+++ src/transforms/v2-to-v3/__fixtures__/api-promise/service-import.input.js
@@ -4,16 +4,3 @@ const client = new DynamoDBClient();
 
 // ExpressionStatement
 client.listTables().promise();
-
-// Promise call on request in variable declarator
-const listTablesRequest = client.listTables();
-listTablesRequest
-  .promise()
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
-
-const listTagsOfResourceRequest = client.listTagsOfResource({ ResourceArn: "STRING_VALUE" });
-listTagsOfResourceRequest
-  .promise()
-  .then((data) => console.log(data))
-  .catch((err) => console.log(err, err.stack));
