+++ src/transforms/v2-to-v3/__fixtures__/api-promise/var-declarator-promise.output.js
@@ -1,0 +1,14 @@
+import { DynamoDB } from "@aws-sdk/client-dynamodb";
+
+const client = new DynamoDB();
+
+// Promise call on request in variable declarator
+const listTablesRequest = client.listTables();
+listTablesRequest
+  .then((data) => console.log(data))
+  .catch((err) => console.log(err, err.stack));
+
+const listTagsOfResourceRequest = client.listTagsOfResource({ ResourceArn: "STRING_VALUE" });
+listTagsOfResourceRequest
+  .then((data) => console.log(data))
+  .catch((err) => console.log(err, err.stack));
